unsupported modules: new logic for SLE15
On SLE15, unsupported modules are only allowed if the workstation
extension is installed (modules in kernel-default-extra). We
do away with the complex old logic and enforce this setting in
10-unsupported-modules.conf even on update.

Enabling unsupported modules for WE is handled by the new
suse-module-tools-we package.

Add more comments in 10-unsupported-modules.conf explaining the
rationale.