modsign-verify: try multiple certificates
The openSUSE Secure Boot CA and and the openSUSE Secure Boot
Signkey have identical issuer and serial number and thus can't be
distinguished simply by comparing these fields.

It is safe to try the signature from the first matching certificate
and continue searching if that fails, as long as we return failure if no
matches are found.