spec file: fix %config(noreplace) behavior
TL;DR: the currently released implementation behaves weirdly on updates
because the installed file /etc/modprobe.d/10-unsupported-modules.conf
is modified during %post on SLE products and thus appears as "user-modified"
to the package management. Fix that by using different defaults for
allow_unsupported_modules: 0 for SLE and 1 for openSUSE.

The current code defaults to "allow_unsupported_modules 1" and changes the
value to 0 in the %post script if a SLE environment is detected and certain
other conditions hold. This means that, in those cases where
allow_unsupported_modules is 0, rpm considers the file "modified". rpm can't
distinguish the case "modified during %post" from "modified by user". That's
unfortunate, because we can't easily determine whether or not the user
modified the file. Together with %config(noreplace), this logic ensures that
/etc/modprobe.d/10-unsupported-modules.conf is *never* replaced by rpm updates
on SLE12.

... well, that's not 100% true either. If (on SLE12) the user edited the file
replacing "allow_unsupported_modules 0" with "allow_unsupported_modules 1",
the checksum of the file would be *correct* again, making rpm think that the
file was unmodified. The %config(noreplace) logic would then replace the file
if it had changed in an update (say the default set to 0). But the %post logic
would _not_ trigger because it does nothing during updates. Thus the
"%config(noreplace)" logic is *inverted*: on systems where the user changed
the file, it's replaced, and on systems where the user kept the default, it's
not.

(This again would not hold if the user also did some other modifications to
the file (add whitespace and whatnot) in which case his modifications would
stay in place... sigh).