Use modules-load.d/sg.conf for sg driver autoloading
It turns out that using an udev rule for loading the sg
driver might be "overkill" on systems with lots of SCSI
devices. udev used to have such a rule until v173, which
has been removed in commit 09637f74. It was recommended
to load sg statically instead.

Therefore we now simply use the sg.conf approach, taken
from the systemd package.