modprobe.conf: add file system blacklist file (fate#326832).
Some file systems are not used on most systems but the kernel
will still happily load the modules for them when presented with
a potentially malicious device like a USB stick.  Many of these
have not been maintained except for in-kernel API compatibility
in many years, have not been audited, and may have security
vulnerabilities.

This patch adds a new 60-filesystem-blacklist.conf that prevents
the kernel from autoloading these modules.