spec file: implement filesystem module blacklisting (jsc#SLE-4085)
All the blacklisting-related logic is in the spec file. The default is
to blacklist all previously mentioned modules. Separate blacklist files
for each file system are placed in /etc/modprobe.d/60-blacklist_fs-$FS.conf.

If any such module is loaded during the installation of the suse-module-tools
package (i.e. listed in /proc/filesystems), the %post script comments out the
blacklist line on the bottom. This will cause the file to appear modified to
rpm, so that future updates won't overwrite it, as the files are marked
%config(noreplace). %config(noreplace) doesn't protect the files from being
modified by the %post script though. Therefore, people who want to keep the
FS blacklisted (even though they sometimes load the module manually) can remove the
"__THIS FILE MAY BE MODIFIED__" marker. That way, even if the suse-module-tools
rpm was updated while the module was loaded, it wouldn't un-blacklist it.