spec file: drop jffs2 and ubifs from blacklist
R. Wenberger, Jan 14, 2019:
"[jffs] is a raw flash filesystem, it does not work on a block device. So there
is no way to have this automatically mounted by plugging in a USB thumb
drive.

Jeff Mahoney, Jan 15, 2019:
"True, and the code doesn't allow loopback mounts. We can drop jffs2."
"... The same applies to ubifs. "