Load fbcon together with virtio_gpu on s390
The framebuffer console is normally built into the kernel, but it's
essentially dead code if there is no video driver. Since System z
does not have graphics hardware, the only available DRM module is
virtio_gpu. This driver is used when running as a KVM guest with a
virtio video adapter. This scenario is the only one where fbcon can
be useful on System z, but then the fbcon module should be loaded
automatically.