weak-modules2: emit "inconsistent" warning only if replacement fails
The purpose of the warning "/lib/modules/$krel is inconsistent" is
to detect a situation where a valid KMP update can't be installed
because some other, unrelated module (or KMP) has unresolved symbols.
A check for unresolved symbols is made before removing the weak
symlinks of the KMP and before trying to install a replacement.
(Reference: bnc#426957, https://build.opensuse.org/request/show/82552).

However, in certain situations, unresolved symbols which were present
before removing the KMP would be fixed later, by installing the
replacement. This happens if some other module depends on symbols provided
by the KMP being updated, the replacement module provides the same symbols,
and the symlinks for the KMP being replaced were already removed before
invoking "wm2 --remove-kmp". The latter can happen if several versions of the
same KMP providing "multiversion(kernel"), built against different kernel
sources, are installed, and are being replaced by a single KMP not providing
"multiversion(kernel)". See bsc#1127155 for details. Now a spurious warning
would be emitted in a healthy situation, irritating users.

This patch attempts to fix this by collecting the unresolved symbols
at the same step as before, but only emitting the warning if possible
replacement candidates for the KMP were available, but couldn't be
installed.