modprobe.conf: use SUSE INITRD for cifs/nls, again
e6f7e30 ("Remove more SUSE INITRD instructions") changed the SUSE
INITRD line for cifs in modprobe.conf to a "softdep pre". But this
would cause nls_utf8 to be always loaded before cifs, which isn't
necessary. We do have a the dracut module for "SUSE INITRD", so
leaving this one in seems leaner.

Fixes: e6f7e30 ("Remove more SUSE INITRD instructions")