spec file: move modprobe and depmod conf files to /lib
Move /etc/modprobe.d to /lib/modprobe.d and /etc/depmod.d to
/lib/depmod.d.

In principle, the "right" location for these directories would be
/usr/lib/modprobe.d and /usr/lib/depmod.d, respectively. But kmod has only
supported this for modprobe.d since recently (Nov. 2020), and doesn't support
it for depmod.d until now (May 2021). Using /lib is safe, and files will be
transparently moved to /usr/lib/modprobe.d with the usr merge.

All modprobe.conf files are moved from /etc/modprobe.d to /lib/modprobe.d.
The override functionality for the filesystem blacklist now works
with the standard "/etc overrides /lib" mechanism. "Automatic" override
files in /etc will be created in %post if the respective file system modules
are loaded while %post is running. This corresponds to the way previous
versions altered the files under /etc during %post. Files under /etc will
only be overwritten if they contain the string "__THIS FILE MAY BE MODIFIED__".
The "%config(noreplace)" attribute of the conf files is removed, as these
are now static, and not meant to be changed any more.

The file 99-local.conf, which was a placeholder for local modifications,
is dropped, because it makes no sense to keep it in /lib.