weak-modules2: unique logic for treating compression
Separate out the stripping of the extension into a general function,
strip_mod_extensions(). Introduce get_initrd_basenames() for determining
the list of module names in the initrd.

NOTE: We can NOT strip the extensions in basenames-$kmp.
This would lead to inconsistent behavior in basenames_are_unique():
For the kernel they were built for, we could have both compressed and
uncompressed modules installed at the same time, but for compatible
kernels, we couldn't have them both symlinked in weak-updates. This
leads to unexpected results if the modules pointed to by the symlinks
are removed.