sysctl kernel-uname_r.conf: separate service units, kernel usr merge
We used to pull in the kernel-specific sysctl settings via an ExecStartPre
directive in systemd-sysctl.service. Create a separate service instead,
which is pulled in by systemd-sysctl. This is more transparent, and allows
more fine-grained control.

In preparation of the kernel /usr merge (bsc#1184804), create two separate
services boot-sysctl. service and kernel-sysctl.service which work with kernel
packages pre- and post- usrmerge, respectively. Only one of them will be
actually started because of a Condition directive.

"boot-sysctl.service" can be dropped when we may assume that no users keep
old kernel packages around that install their sysctl settings under /boot.