weak-modules2: handle kernel modules in /usr/lib/modules
On the input side the script needs to handle /usr/lib/modules. When
checking in the file system it does not necessarily have to as /lib
link to /usr/lib. So as long as the script potentially has to deal with
both locations it's ok to poke around in /lib to keep the code simple.
As soon as we only need to handle /usr/lib all occurrences can be
replaced.