weak-modules2: accept modules under /usr/lib/modules on stdin
This logic relies on the assumption that /lib/modules always resolves
to /usr/lib/modules. wm2 continues to work with /lib/modules only,
internally. All symlinks created will point to files under /lib/modules.

Thus the only point in the code where we need to account for /usr is
when we read the list of modules from stdin, and there we simply strip
off the /usr prefix, knowing that the result will still resolve to the
same file.