weak-modules2: skip "test" call to has_unresolved_symbols
The first call is supposed to check whether the kernel at hand
is in an inconsistent state before adding the KMP. But this is
a very rare situation; has_unresolved_symbols() is a slow operation (calling
depmod) and should thus not be called without good reason.

If adding the kmp failed and OPT_DEBUG=1, we carry out this operation
to provide a clue to users.