weak-modules2: honor INITRD_IN_POSTTRANS for kernel packages
With this patch, if the $INITRD_IN_POSTTRANS environment variable
is set in weak-modules2's environment, it will postpone the rebuilding
of the initramfs to the %posttrans stage if possible. Also, for kernel
subpackages like -extra or -optional, the initrd rebuild will only
be triggered if it's necessary (i.e. if the current initrd contains
at least one of the modules in the package).