inkmp-script(postun): don't pass  existing files to weak-modules2
It can happen that %postun encounters files in the list of removed
modules that still exist, because another kernel package has replaced
them. weak-modules2 currently does nothing with the list of modules
passed to it in this code path, but in case it will do so some time
in the future, pass the correct list.