weak-modules2: skip initrd rebuild in remove_kernel_modules
We are removing modules here which might be in the initrd.
We can derive kernel-modules and kernel-basenames like we
do in add_kmp_modules. Why don't we do it?

The common case by far is that kernel subpackages are removed
or updated together with the base package. In that case, rebuilding
the initrd in %postun of the subpackage is just a waste, as the
kernel will be uninstalled anyway.

In %postun we can't figure out whether or not the base package
will also be uninstalled. If it's not, in the worst case, some modules
will be kept in initramfs that are no longer installed on the system
itself. That shouldn't hurt. By not rebuilding the initrd here,
we speed up the common case.

We have to call depmod though, in order to make sure that the module
dependency database doesn't hold stale references to uninstalled modules.