Keep br_netfilter softdep on SLE15 (GA)
Revert "spec: use br_netfilter softdep only for SLE12"
This reverts commit e75342eb6391b56384bc5159dc48ed66bd878c8e.

The softdep had been removed in Leap 15.1 and later, originally.
With "closing the leap gap", we decided to remove it in SLE15-SP2
and later only, and did the same for SLE15-SP1, which was still
in general support at that time (mid 2020). However SLE15 (GA)
was in LTSS, and still is. I didn't expect to have to release
another suse-module-tools version for SLE15 (GA). But it must
be done now in order to provide the kernel scriptlets (bsc#1189841).
"closing the leap gap" is no issue for SLE15 (GA).

I don't think changing the behavior wrt br_netfilter makes sense
so deeply in LTSS. Therefore I'm reverting this change just for
SLE15 (GA).