rpm-script: force copy kernel to /boot (boo#1194501)
kernel uname -r is actually without the OBS rebuild counter. It
could happen that a kernel package gets rebuilt, resulting in a new
kernel with same uname -r that cannot load modules from the previous
build with the same version. So the kernel in /boot needs to be
updated always, even if a file with same name already exists there.