modprobe.conf: s390x: remove softdep on fbcon
Since 6104c37094e7 ("fbcon: Make fbcon a built-time depency for fbdev",
upstream 4.14), fbcon is not a separate module any more. The framebuffer
console is enabled at compile time via CONFIG_FRAMEBUFFER_CONSOLE=y,
which is enabled on all SUSE kernels.

modprobe handles this gracefully as long as only virtio_gpu is loaded; it just
skips the softdep. But the suse-initrd dracut module throws an error trying to
copy the fbcon module into the initramfs (bsc#1207853).

Drop the dependency on fbcon.