Revert "rpm-script: detect usrmerged automatically"
This reverts commit 3a6e48605b1dc4637d51c71ca788ed4fd46e5c42.

The [ -L /lib ] sounds less reliable than the kernel to telling the
scripts when the modules are placed in /usr.

The kernel always knows where it placed the modules.

On the other hand, the /lib symlink only exists when a specific package
is installed. This may not be the case due to package installation
ordering or in special environments not directly managed by rpm.