split off regenerate-initrd-posttrans in the kernel-scriptlets package
Users or packages may want to override the regenerate-initrd-posttrans script.
Split it off into the kernel-scriptlets subpackage (bsc#1213459).

Rationale: systemd-boot changes the way the boot loader works so massively
that we can't easily integrate it into the current (perl-)bootloader
abstraction. Booting from btrfs snapshots makes this even more
difficult, the sdbootutil package is being developed for dealing with that
issue. The file name of the initramfs image is related to its checksum,
thus can't be determined before the image has been generated.

The kernel scriptlets have to be changed for sd-boot integration anyway.
Therefore, move r-i-p into that subpackage for now. All scriptlets call
r-i-p only if it is installed, anyway, therefore r-i-p need not be required
by any package. `/run/regenerate-initrd` is still supported as an indicator
that the initrd will need to be rebuilt.

Signed-off-by: Martin Wilck <mwilck@suse.com>