Fix kernel test in regenerate-initrd-posttrans (bsc#1214877).
commit 60a2a14f978e ("weak-modules2: only use kernel version under
/run/regenerate-initrd") changes the API of regenerate-initrd-posttrans
to only take the kernel version, not the image name.

It also checks that the module directory exists, not the kernel image.

This is both broken.

While the image is removed with the kernel the module directory may
contain KMPs built against this kernel version and remain.

Anyway, at least revert the API change, there seem to be some users
outside of suse-module-tools.

The problem with the remaining kernel module directory should be
resolved by moving out-of-tree KMPs outside of the kernel module
directory.

Fixes: 60a2a14f978e ("weak-modules2: only use kernel version under /run/regenerate-initrd")
Signed-off-by: Michal Suchanek <msuchanek@suse.de>