rpm-script: fix handling of /boot/vmlinuz and /boot/initrd
The symlinks /boot/vmlinuz and /boot/initrd should only be created
if their targets exists (which may not be the case e.g. if dracut
had failed to create the initramfs). This requires that we create
these symlinks late in the installation process, at the same
time when we create the bootloader entries.