rpm-script: create dangling symlinks in kiwi environments
In image build environments such as kiwi (more precisely, if /.buildenv
exists), dracut is not invoked by weak-modules2. That means that the logic
of 19c92c5 ("rpm-script: fix handling of /boot/vmlinuz and /boot/initrd"),
which tries to avoid creating dangling symlinks, doesn't work.

While this should eventually be fixed in kiwi, we can't wait for that to happen.
In such cases, fall back to creating a dangling symlink, hoping that kiwi will
run dracut later, creating the symlink's target.

Fixes: 19c92c5 ("rpm-script: fix handling of /boot/vmlinuz and /boot/initrd")
Signed-off-by: Martin Wilck <mwilck@suse.com>