Find update candiate links across all module subdirectories
If module subdirectories are versioned, we need to search a bit harder
for potential update candiates as the name of the module subdirectory
differs.
We therefore search all weak-updates subdirectories of the $krel kernel
for a module module matching the same name (taking into accout different
compressions).
Should we find multiple modules that match (which should not happen!)
we flag this but continue regardless. This way, the first element in
the list will be taken as the potential update candidate.

Signed-off-by: Egbert Eich <eich@suse.com>