weak_modules2: symlink_to_module: allow almost arbitrary paths
While the symlink will always be under
/usr/lib/modules/$(uname -r)/weak-modules, the module itself can
be installed (almost) anywhere. For backward compatibility, if the
module is installed in kernel directory like traditional KMPs,
use the relative path of the module wrt the kernel directory
as relative path of the symlink under weak-updates.
Otherwise, if the path ends in /updates/, use that as relative path.
Otherwise, simply install the symlink under weak-updates directly.

For safety, print an error message if the module path isn't absolute, and
deal with empty return value in callers.

Signed-off-by: Martin Wilck <mwilck@suse.com>