weak-modules2: distinguish available modules
With the "arbitrary directory" approach, a module that is built
for a given kernel isn't necessarily installed in that kernel's
module directory (aka "available"). Use the file system flag
krel-$kmp-available to indicate if the modules of this KMP are
available. If not, symlinks must be created and removed also
for the kernel the module was built against.

Signed-off-by: Martin Wilck <mwilck@suse.com>